Update Scala.js Version + Init API
