+++ index.html
@@ -10,7 +10,7 @@
     <script type="text/javascript" src="./target/scala-2.11/p5-scalajs-fastopt.js"></script>
     <!-- Run tutorial.webapp.TutorialApp -->
     <script type="text/javascript">
-      p5_scalajs.demo.Demo().main();
+      p5.demo.DemoScala().main();
     </script>
   </body>
 </html>
